MainViewModel-completed

+++ app/src/main/java/com/messer_amd/listacumparaturilor/domain/ShopItem.kt
@@ -2,7 +2,7 @@ package com.messer_amd.listacumparaturilor.domain
 
 data class ShopItem(
     val name: String,
-    val count: Float,
+    val count: Int,
     val enabled: Boolean,
     var id: Int = UNDEFINED_ID
 )
